extensions: add APA extension
using this extension allows for activation of the APA armbian package
archive and installation of the provided binary packages at image
creation time from within the Armbian Build Framework.

The goal of APA is to simplify maintenance of Armbian Core by moving
dependency logic and other packaging information into a separate space
and handle that meta-data with the proper distribution-creation and
-publication tools.  Currently, much of this is done with fragile bash
scripting in Armbian Core instead.

https://github.com/armbian/apa